Build sigrok with C++ bindings, libusb and libftdi and then upload it
To enable C++ bindings: libsigcpp was restricted to the version 2
(otherwise glibmm has unmet requirements). It's not needed in the 'run'
section.

To enable libusb: configure didn't find it when it was in the 'build'
requirements so it was placed as the 'host' requirement. It was
restricted to the '1.0.20' version to use litex-hub's version of
'libusb'.

With the 'channel_priority: strict' however it has to be the channel
with the highest priority. conda-forge's 'libusb' uses 'libudev0' which
is problematic but this channel is needed for other packages.

Also 'libftdi' was added to make a few more hardware drivers available.